services/ui_iot/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
asyncpg==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.0
python-dateutil
python-jose[cryptography]==3.3.0
pytest-asyncio>=0.21.0
pytest>=7.4.0
pytest-cov>=4.1.0
respx>=0.20.0
playwright>=1.40.0
pytest-playwright>=0.4.0
pysnmp-lextudio>=5.0.0,<6.0
pyasn1>=0.4.8,<0.6.0
aiosmtplib>=3.0.0
email-validator
paho-mqtt>=2.0.0